+++ docs/ChangeLog.rst
@@ -2,6 +2,9 @@
 Release Notes
 =============
 
+Version 1.0 (in development)
+============================
+
 Version 0.9
 ===========